Further effort porting LL curses interface to DE.
LLSupport/ll-sequence.c
LLSupport/lloptions.c
LLSupport/Includes/ll-sequence.h
LLSupport/Includes/lloptions.h
Lifelines/signals.c
Lifelines/version.c
Lifelines/Includes/signals.h
Lifelines/Includes/version.h
    New files.

DataTypes/Includes/hashtable.h
    struct HashTable: new field: refcount
    New decls: addrefHashTable, releaseHashTable

DataTypes/hashtable.c
    createHashTable: initialize count, refcount.
    new functions: addrefHashTable, releaseHashTable.

Database/Includes/refnindex.h
    struct RefnIndexEl: make refn and key const.
    createRefnIndexEl: make key const

Database/keylist.c
    cast strlen return value to int.

Database/refnindex.c
    createRefnIndexEl: make key const.

Interp/Includes/sequence.h
Interp/sequence.c
    appendToSequence: make key and name const.

LLSupport/Includes/de-strings.h
    allocsubbytes: new function.

LLSupport/Includes/ll-porting.h
    release_table, addref_table: map to releaseHashTable, addrefHashTable

LLSupport/Includes/locales.h
    Convert LL types to DE equivalents.

LLSupport/Makefile.am
    add files: ll-sequence.c, lloptions.c

LLSupport/ll-str.c
    New function: allocsubbytes

LLSupport/write.c
    #if 0 swrite_node, swrite_nodes, node_to_string -- DE provides them now

Lifelines/Makefile.am
    New files: signals.c, version.c

Lifelines/browse.c
    Include ll-sequence.h; call getAllSources, getAllEvents, getAllOthers

Lifelines/main.c
    Don't call more functions: init_arch, init_stdlib, init_interpreter,
    shutdown_interpreter.  More changes to come.

Lifelines/screen.c
    Don't define edit_user_options (maybe someday). #if out call to it.
    call get_deadends_version, not get_lifelines_version

Lifelines/show.c
    add_child_line: second argument is now a GNode *, not a RECORD.

Python/Includes/porting.h
    Comment out getlloptstr macros.

Python/python-to-c.c
    Adjust comment; this is DeadEnds, not LifeLines.

Python/run.c
    New includes: locales.h, lloptions.h

Utils/utils.c
    Cast strlen return value

configure.ac
    DEADENDS_CONFIG_FILE, not LINES_CONFIG_FILE; adjust filename as well.

+++ Database/keylist.c
@@ -21,7 +21,7 @@ static int keyCompare(Word a, Word b)
 	// (This duplicates code from sequence.c. Should be cleaned up.)
 	ASSERT(strlen(keya) > 2 && strlen(keyb) > 2);
 	if (strlen(keya) != strlen(keyb)) return (int) (strlen(a) - strlen(b));
-	for (int i = 0; i < strlen(keya); i++) {
+	for (int i = 0; i < (int)strlen(keya); i++) {
 		if (keya[i] != keyb[i]) return keya[i] - keyb[i];
 	}
 	return 0;
@@ -42,7 +42,7 @@ static int rootKeyCompare(Word a, Word b) {
 	// (This duplicates code from sequence.c. Should be cleaned up.)
 	ASSERT(strlen(keya) > 1 && strlen(keyb) > 1);
 	if (strlen(keya) != strlen(keyb)) return (int) (strlen(a) - strlen(b));
-	for (int i = 0; i < strlen(keya); i++) {
+	for (int i = 0; i < (int)strlen(keya); i++) {
 		if (keya[i] != keyb[i]) return keya[i] - keyb[i];
 	}
 	return 0;
